Hub and groups
The following uses task management to handle subscribers to a hub. It
is up to the subscribers to consume messages from the hub and not the
hub to iterate over the messages dispatching them. This allows for a
more spoke(-d) type of hub, where the consuming can be distributed and
prevents the hub from total blocking.

The group type allows for fairness over the consumption of the messages,
but doesn't guarantee it, only different timings dictate this, no
schedule priorities in the hub do this.

Collapsing of messages is possible when backed up, but in reality is
never worth the time.