Backoff tasks
The following implements a backoff algorithm when running tasks. If a
task returns a ErrBackoff it allows the scheduler to unknow that a
backoff is wanted. It is then up to the function to decide what sort of
back off is required.

This is really useful when you don't want to be spin locking all the
time.