Non-consumed backoff
The following changes how backoff is implemented at the subscriber level.
Instead of always returning a task.ErrBackoff when the queue is empty, we
now check to see if there has been any consumption of the queue. If the
queue consumed anything then we should just return nil. The theory
being that high activity is followed by long pauses of inactivity. This
fundamental idea is then, that we will be very fast at consuming messages
when there are messages, but if there aren't any, we just slowly backoff
enough for other things to run.

Using the metrics hook, shows us for _our_ workload this is good enough.
Others will probably want to make that choice themselves.

In the future a runtime scheduler could be used to determine how to run
the loop (subcriber.run), but for now and for our purposes this is fine.

There is always a max backoff that prevents the wait of the loop going
to long before checking a message.